% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rid_zero.R
\name{get_rid_zero}
\alias{get_rid_zero}
\title{eliminate empty rows}
\usage{
get_rid_zero(r)
}
\arguments{
\item{r}{matrix}
}
\value{
new matrix remaining rows with values
}
\description{
eliminate empty rows
}
\examples{
set.seed(123)
x = matrix(rpois(100, .1), 10, 10)
get_rid_zero(x)
}
