% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_0s_by_col.R
\name{rm_0s_by_col}
\alias{rm_0s_by_col}
\title{remove zeros vertically /w colsums}
\usage{
rm_0s_by_col(col)
}
\arguments{
\item{col}{matrix column}
}
\value{
a matrix with occupied columns
}
\description{
remove zeros vertically /w colsums
}
\examples{
set.seed(123)
x = matrix(rpois(100, .1), 10, 10)
rm_0s_by_col(x)
}
