% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_0s_by_row.R
\name{rm_0s_by_row}
\alias{rm_0s_by_row}
\title{removes zeros from matricies by row}
\usage{
rm_0s_by_row(r)
}
\arguments{
\item{r}{matrix}
}
\value{
remaining rows with values, matrix
}
\description{
removes zeros from matricies by row
}
\examples{
set.seed(123)
x = matrix(rpois(100, .1), 10, 10)
rm_0s_by_row(x)
}
